+++ jetty-core/jetty-http3/jetty-http3-common/src/main/java/org/eclipse/jetty/http3/generator/DataGenerator.java
@@ -35,16 +35,16 @@ public class DataGenerator extends FrameGenerator
     }
 
     @Override
-    public int generate(ByteBufferPool.Accumulator accumulator, long streamId, Frame frame, Consumer<Throwable> fail)
+    public long generate(ByteBufferPool.Accumulator accumulator, long streamId, Frame frame, Consumer<Throwable> fail)
     {
         DataFrame dataFrame = (DataFrame)frame;
         return generateDataFrame(accumulator, dataFrame);
     }
 
-    private int generateDataFrame(ByteBufferPool.Accumulator accumulator, DataFrame frame)
+    private long generateDataFrame(ByteBufferPool.Accumulator accumulator, DataFrame frame)
     {
         ByteBuffer data = frame.getByteBuffer();
-        int dataLength = data.remaining();
+        long dataLength = data.remaining();
         int headerLength = VarLenInt.length(FrameType.DATA.type()) + VarLenInt.length(dataLength);
         RetainableByteBuffer header = getByteBufferPool().acquire(headerLength, useDirectByteBuffers);
         ByteBuffer byteBuffer = header.getByteBuffer();
