+++ jetty-core/jetty-http3/jetty-http3-common/src/main/java/org/eclipse/jetty/http3/generator/HeadersGenerator.java
@@ -39,13 +39,13 @@ public class HeadersGenerator extends FrameGenerator
     }
 
     @Override
-    public int generate(ByteBufferPool.Accumulator accumulator, long streamId, Frame frame, Consumer<Throwable> fail)
+    public long generate(ByteBufferPool.Accumulator accumulator, long streamId, Frame frame, Consumer<Throwable> fail)
     {
         HeadersFrame headersFrame = (HeadersFrame)frame;
         return generateHeadersFrame(accumulator, streamId, headersFrame, fail);
     }
 
-    private int generateHeadersFrame(ByteBufferPool.Accumulator accumulator, long streamId, HeadersFrame frame, Consumer<Throwable> fail)
+    private long generateHeadersFrame(ByteBufferPool.Accumulator accumulator, long streamId, HeadersFrame frame, Consumer<Throwable> fail)
     {
         RetainableByteBuffer buffer;
         // Reserve initial bytes for the frame header bytes.
@@ -64,7 +64,7 @@ public class HeadersGenerator extends FrameGenerator
             encoder.encode(byteBuffer, streamId, frame.getMetaData());
             byteBuffer.flip();
             byteBuffer.position(maxHeaderLength);
-            int dataLength = buffer.remaining();
+            long dataLength = buffer.remaining();
             int headerLength = frameTypeLength + VarLenInt.length(dataLength);
             int position = byteBuffer.position() - headerLength;
             byteBuffer.position(position);
